// src/interfaces/clientsInterface.tsx
// Generated by https://quicktype.io

export interface ClientResponse {
  id: string;
  fullName: string;
  email: string;
  phone: string;
  photo: string;
  state: number;
  createdAt: string;
  updatedAt: null;
  deletedAt: null;
  account: Account;
  app: App;
}

export interface Account {
  id: string;
  clientId: string;
  balance: string;
  credit: string;
  state: number;
  createdAt: string;
  updatedAt: null;
  deletedAt: null;
}

export interface App {
  id: string;
  clientId: string;
  appColor: string;
  createdAt: string;
  updatedAt: null;
}